requirements.txt
# requirements.txt
uagents
hyperon
google-generativeai
httpx
openai
python-dotenv
fastapi
uvicorn
jinja2
